requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
mysql-connector-python==8.0.27
